+++ packages/ts-components/src/helpers/layout-size/layoutSize.ts
@@ -1,0 +1,12 @@
+export enum Layout {
+  Standard = '4043',
+  Wide = '4042'
+}
+
+export const isStandard = (layoutSize: Layout) => {
+  return layoutSize === Layout.Standard;
+};
+
+export const isWide = (layoutSize: Layout) => {
+  return layoutSize === Layout.Wide;
+};