+++ packages/ts-components/src/index.ts
@@ -5,6 +5,7 @@ export { GalleryCarousel } from './components/carousel/GalleryCarousel';
 export {
   InfoCardBulletPoints
 } from './components/in-article-info-card-bulletpoints/InfoCardBulletPoints';
+export { BigNumbers } from './components/in-article-big-numbers/BigNumbers';
 
 // Newsletter Components
 export {